+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - [x] Add textbox for arguments
 - [x] Better sorting of package names
+- [x] Version number completion for NuGet
 - [ ] Implement better auto-completion mechanism
 
 Features that have a checkmark are complete and available for