+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - [x] Add textbox for arguments
 - [x] Better sorting of package names
 - [x] Version number completion for NuGet
+- [x] Support for Typings (#7)
 - [ ] Implement better auto-completion mechanism
 
 Features that have a checkmark are complete and available for