+++ CHANGELOG.md
@@ -11,6 +11,12 @@ download in the
 These are the changes to each version that has been released
 on the official Visual Studio extension gallery.
 
+## 1.7
+
+**2016-05-02**
+
+- [x] Support for "VS15"
+
 ## 1.6
 
 **2016-03-14**